fix pvc cluster ingestion, pvallocaiton json marshalling
