+++ pkg/kubecost/allocation_test.go
@@ -379,78 +379,6 @@ func TestAllocation_AddDifferentController(t *testing.T) {
 
 }
 
-func TestAllocation_MarshalJSON(t *testing.T) {
-	start := time.Date(2021, time.January, 1, 0, 0, 0, 0, time.UTC)
-	end := time.Date(2021, time.January, 2, 0, 0, 0, 0, time.UTC)
-	hrs := 24.0
-
-	gib := 1024.0 * 1024.0 * 1024.0
-
-	cpuPrice := 0.02
-	gpuPrice := 2.00
-	ramPrice := 0.01
-	pvPrice := 0.00005
-
-	before := &Allocation{
-		Name: "cluster1/namespace1/node1/pod1/container1",
-		Properties: &AllocationProperties{
-			Cluster:   "cluster1",
-			Node:      "node1",
-			Namespace: "namespace1",
-			Pod:       "pod1",
-			Container: "container1",
-		},
-		Window:                NewWindow(&start, &end),
-		Start:                 start,
-		End:                   end,
-		CPUCoreHours:          2.0 * hrs,
-		CPUCoreRequestAverage: 2.0,
-		CPUCoreUsageAverage:   1.0,
-		CPUCost:               2.0 * hrs * cpuPrice,
-		CPUCostAdjustment:     3.0,
-		GPUHours:              1.0 * hrs,
-		GPUCost:               1.0 * hrs * gpuPrice,
-		GPUCostAdjustment:     2.0,
-		NetworkCost:           0.05,
-		LoadBalancerCost:      0.02,
-		PVs: PVAllocations{
-			disk: {
-				ByteHours: 100.0 * gib * hrs,
-				Cost:      100.0 * hrs * pvPrice,
-			},
-		},
-		PVCostAdjustment:       4.0,
-		RAMByteHours:           8.0 * gib * hrs,
-		RAMBytesRequestAverage: 8.0 * gib,
-		RAMBytesUsageAverage:   4.0 * gib,
-		RAMCost:                8.0 * hrs * ramPrice,
-		RAMCostAdjustment:      1.0,
-		SharedCost:             2.00,
-		ExternalCost:           1.00,
-		RawAllocationOnly:      &RawAllocationOnlyData{},
-	}
-
-	data, err := json.Marshal(before)
-	if err != nil {
-		t.Fatalf("Allocation.MarshalJSON: unexpected error: %s", err)
-	}
-
-	after := &Allocation{}
-	err = json.Unmarshal(data, after)
-	if err != nil {
-		t.Fatalf("Allocation.UnmarshalJSON: unexpected error: %s", err)
-	}
-
-	// TODO:CLEANUP fix json marshaling of Window so that all of this works.
-	// In the meantime, just set the Window so that we can test the rest.
-	after.Window = before.Window.Clone()
-	// TODO Sean: fix JSON marshaling of PVs
-	after.PVs = before.PVs
-	if !after.Equal(before) {
-		t.Fatalf("Allocation.MarshalJSON: before and after are not equal")
-	}
-}
-
 func TestAllocationSet_generateKey(t *testing.T) {
 	var alloc *Allocation
 	var key string
